chore(Catalog): Clean catalog folder prior build
Currently, the `catalog` folder is kept between builds, and that's
problematic because it keeps old files that are no longer referenced in
any index, making the bundle bigger than it needs to be.

This doesn't have an impact in runtime because only the needed files are
loaded, but yet it's not an efficient use of the storage.

The fix is to clean the `catalog` folder before building so it only
contains the right files.